stamp: who publishes a completion word is the rail's answer, not a cfg's
`write_stamp` and the root drain each carried the same two-armed `cfg`: a
Vulkan block that could attach the stamp word to a GPU submission, and a
`not(backend-vulkan)` line that blocked and wrote the word on the CPU. The
Vulkan block reached three engine entry points and its own three-state
`StampOrder` directly from the drain, so the drain's control flow was the rail's
control flow.

`Backend::order_completion_stamp` names the obligation instead: before it
returns, everything this device owes guest RAM for the completed work has
landed — the guest's fence moving is what frees it — and the answer says only
*who writes the word*. `StampOrdering::Queued` (the rail took it, and raises the
interrupt itself) or `Settled` (the caller writes it). The default is the
blocking settle, which is exactly what the Metal arm did and is the correct
answer for any rail that cannot express the ordering as a submission.

`StampOrder`'s third state stays where it is load-bearing. `CpuReady` ("nothing
was outstanding") and `Declined` ("the async route failed") must not be
confused — a fallback that re-read global debt would wait on work another thread
armed *after* this stamp — but both reach the drain as the same instruction, so
the distinction lives inside `runtime/drain/vulkan.rs` where the settle either
happens or provably need not, and collapses at the seam.

`note_stamp_direction` becomes `pub(super)` (the rail takes the reading before
enqueueing, since the completion thread owns the next write), and the
`StampOrder` unit test moves into the rail with the type.

The comment above the call in `write_stamp` kept its evidence — the 810-of-810
page-set guard, the `gpu_stamps=0` beside a matching `readback_split` — and
dropped the sentences describing quiesces that no longer sit below it.

No behaviour change: each arm executes the body it compiled before. Verified on
both arms — feature matrix, clippy `-D warnings`, full suites serially (metal
1326, vulkan 2269, support 351, unchanged).
